requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
google-generativeai==0.3.2
gTTS==2.5.0
pydub==0.25.1
python-dotenv==1.0.0
numpy==1.26.4
SpeechRecognition==3.10.0